Added Personality Script for Facial GUI
Added the personality.py script for facial GUI display. Features eyes with pupils that move and change corresponding with current robot activity and motion of direction. Features a subtitles option that displays the text being spoken. Needs to have the text being spoken sent as input to the script for use in subtitles.