Uploaded Test, but Actually Working, New Driver
Uploaded guiTests.py, the new "driver" we'll be using. For some reason the nearly identical intended new driver won't run the server and gui in parallel, but this will. Currently testing for method call comparisons, need to add in client later. To run, run this script only on the robot.