Uploaded the final project script
Uploaded the final project autonomous movement script (project12.py) merged with the ai text to speech and script, along with the face GUI. Modified movement values slightly and fixed a small bug or two for turning angles. Needs to be tested, but should be the near final form of the complete package. The bot will need to have all relevant images and packages moved to the new folder to run.